+++ src/components/dialog/dialog.stories.js
@@ -1,6 +1,5 @@
 import React, { useRef, useState } from 'react';
 import { addStoryInGroup, MID_LEVEL_BLOCKS } from '../../../.storybook/utils';
-import { IconWarningBadgedMediumOutline } from '@teamleader/ui-icons';
 import { Box, Button, Dialog, TextBody } from '../../index';
 import Input from '../input';
 import Label from '../label';
@@ -30,7 +29,6 @@ export const DefaultStory = (args) => {
       <Button onClick={toggleDialog} label="Open a dialog" />
       <Dialog
         {...args}
-        headerIcon={<IconWarningBadgedMediumOutline />}
         active={active}
         onCloseClick={toggleDialog}
         onEscKeyDown={toggleDialog}
@@ -45,6 +43,11 @@ export const DefaultStory = (args) => {
 };
 
 DefaultStory.args = {
+  leftAction: {
+    label: 'Remove',
+    level: 'destructive',
+    onClick: () => console.log('leftAction.onClick'),
+  },
   primaryAction: {
     label: 'Confirm',
     onClick: () => console.log('primaryAction.onClick'),